backend/app/requirements.txt
asgiref==3.8.1
Django==5.2.1
sqlparse==0.5.3
python-dotenv==1.1.0
psycopg2-binary==2.9.10
argon2-cffi==23.1.0
